fix: changed abs to to relative imports - moved diagram into a 'diagram_utils' package
